Refactor the genome
With this refactorization, the genome becomes modular. It is now trivial
to add new traits with any number of genes per trait. The current use of
the genome remains the same, but the next part of the refactor will
involve moving all scaling functionality (the old "getters") out of the
genome, scaling at the point of usage. With that done, the interface of
the genome will only be the getTraitSum/List/Avg(String) methods.

Because iterators are nice, I've moved from a float[] to an
ArrayList<Float> (not a FloatList because the ArrayList is more usable).

Magic numbers are, for the most part, gone. Instead, traits are
referenced by name, and when a genome is first instantiated, the offsets
for each trait are calculated and hidden from the user.