Refactor the genome (again)
Although I wanted to use an enum for the traits, Processing does not get
along with enums. Ah well.

Replaced HashMap<String, trait> with an inner Trait class, and an
instantiation of it for each named trait. Since it sits inside the
Genome, the total number of genes (and therefore each trait's starting
index) is calculated by the Trait ctor. This is lean, mean, and clean,
such that adding a new trait is as trivial as adding a new instance to
the Genome. The ctor takes care of "inserting" its loci into the genome.

This also means that trait lookups are as simple as, e.g.,
Density.sum() (if inside Genome), and otherwise g.Density.sum() (given
the Genome g).

Since all the traits are known before compilation time, a HashMap was
ill-fit because we have a perfect hash, the traits' names. (Hence why an
enum would fit so nicely.) Although having an instance of each trait in
each instance of the Genome incurs some overheard, it is most certainly
not our bottleneck, and the convenience and familiarity of the approach
outweights this concern.

ArrayList<Float> is replaced with FloatList since it has sublist and sum
methods, and supports iterators.

Following consistent class/method naming conventions (title case), such
that the genome array no longer needs to be prefixed with "the".

Ran whitespace-cleanup in creature.pde and genome.pde.

Addressed mutation bug introduced in last commit because Java iteration
is not backed by the data structure; i.e. for (float f : genome) { //
modifying f does not modify the value in the array }. Had to use
set/get. Oh Java.