Minor change to armor and mutation
Mutation has only been called when the population class made a new
creature from offspring, but not when the population class made a new
creature from scratch. I changed it from population calling creature's
mutation function which calls genome's mutation function, to genome
calling its own mutation function but only in the case where a genome
is created from offspring, not from scratch. This is a common
sense-simplification that does not affect the game in any way. I also
changed the way armor is stored in the genome to make it simpler. It's
now stored like the segment points were, eliminating the need for an
array of Traits.