Clean up creature a bit
While the branch and PR count was down I figured it was a good time to
clean this up a bit.

Much of the same areas of creature will be subject to refactoring of
scaling functions currently in the Genome as soon as this gets merged.

It seems the dominant project style is pretty close to:
[Java Programming Style Guidelines](http://geosoft.no/development/javastyle.html)

I propose we call it the standard just to make things easy.