Fix gametes bug
I encountered a NullPointerException during testing (usually have the
second generation), and debugging proved it to be `gametes[0]`.

After inspecting the declaration, I determined that the
`ArrayList<Genome>` was being used incorrectly, and this is the fix.

This does not change the intention of the code (to collect gametes), but
it raises the question: should we clear the previous generations gametes
for each new one, or continue to allow all gametes to be reused?