Use getGametes() for broadcast spawning
With broadcast spawning, selection does not take place; instead the
gamete pool is filled proportionally to creature's fitnesses, and random
gametes are drawn to creature new creatures.

This means that tournament selection was removed.

This also means that choosing parents (via tournament selection) is
deprecated. The notion of selected and compatible parents is disjoint
from the process that is broadcast spawning.

The speciation and reproduction teams need to determine how
compatibility between two gametes will be handled. For now it's been
removed so that the first step of sexual reproduction can happen.

After the gamete pool is created and the surviving members of the wave
have been saved, the new generation is filled with sexually reproduced
children, with two haploid gametes combining into one diploid child.