Fix #9; kill bodies only on death
The root cause of bug #9 was that box2d bodies were being killed upon
death of the creature (on an update, when health <= 0, alive is set to
false *and* and the body is destroyed), and then being re-killed during
the construction of the next generation (all !alive creatures were being
box2d killed, again).

The jbox2d library destroyBody() function looks like this:

``` java
  public void destroyBody(Body body) {
    assert (m_bodyCount > 0);
    assert (isLocked() == false);
    if (isLocked()) {
      return;
    }
  ...
  }
```

And it was the body count assertion that was failing because of the
double-killing we were doing.