Altitude & Water based on Perlin noise
The altitude of each tile is generated based on Perlin noise, which is a
built in function noise(). Then, using some terrain manipulation
techniques the newly generated altitude map is modified to better
represent terrain found in nature. This mostly includes smoothing the
peaks of the noise function so there aren't bizarre single pixel points
of land, but rather "hills" of land.

Liquid bodies are now calculated based on tile altitude. If a tile is
below some random threshold, it becomes a liquid tile and is shaded
accordingly.