Refactor for "static" traits in Genome
Instead of using static in Processing, since the preprocessor inserts
everything into a single instance of PApplet, moving the traits outside
the genome (as "globals") initializes them only once.

Traits can now be queried on a per-chromosome basis.

The interface for genome values is now closer to our original plan.