Renamed appendages to feelers
From very early on, appendages were to be the creatures' legs and arms
and used for primarily locomotion. These sensory mechanisms shouldn't
share the same name, as that just gets confusing. Henceforth they are
known as feelers, or whatever you guys want them to be called as long
as it's not appendages.