fixed game slow down
Slow down was due to sound threads not stopping.  This is because the
AudioPlayers were never closed.  Fixed this by adding a 2sec delay and
a.close to the Sounds run() function.

Sounds should still play fine as long as they're no longer than 2
seconds (changeable) and no longer slow down the game.