sounds now use keys
Sounds file paths and their lengths have been put into a table now with
each sound having it's own row.  They also have key names associated
with them which is the string that gets sent to the PlaySounds function.

This allows our code to be a little more descriptive as to which sounds
we're playing without needing to send the entire path every time.

I've also added a test case to make sure that the files actually exist
and to print errors if they do not.